requirements.txt
streamlit==1.35.0
sentence-transformers==2.6.1
faiss-cpu==1.8.0
ollama==0.1.34
numpy==1.26.4
torch==2.2.2
tqdm==4.66.2
scikit-learn==1.4.2
langchain==0.1.14
pdfminer.six==20221105
